Use macros to DRY auth tests
