+++ test/test_helper.exs
@@ -6,3 +6,5 @@ ExUnit.start(exclude: [:system])
 Enum.each files, fn(file) ->
   Code.require_file "helper/#{file}", __DIR__
 end
+
+Code.require_file "fennec/udp/auth_template.ex", __DIR__
